pyproject.toml
[project]
name = "deepfer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiortc>=1.9.0",
    "bcrypt>=4.2.0",
    "dlib>=19.24.6",
    "fastapi-socketio>=0.0.10",
    "fastapi[standard]>=0.114.0",
    "matplotlib>=3.9.2",
    "opencv-python>=4.10.0.84",
    "pandas>=2.2.2",
    "pyjwt>=2.9.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.4",
]
